:art: rename wallet import
