+++ app/tests/e2e/issuer/did_sov/test_v2_ld.py
@@ -12,7 +12,7 @@ from assertpy import assert_that
 from app.models.issuer import SendCredential
 from app.routes.issuer import router as issuer_router
 from app.routes.oob import router as oob_router
-from app.routes.wallet import router as wallet_router
+from app.routes.wallet.dids import router as wallet_router
 from app.tests.util.ecosystem_connections import FaberAliceConnect
 from app.tests.util.webhooks import check_webhook_state
 from shared import RichAsyncClient
